refactor(scripts): split override-nextcloud-styles into CLI and function
